Update NSWindow.m
Ensured that all rects, points, and sizes that are sent to back are in pixels that it will understand. The base view is no longer transformed to ensure that all the convertTo… view methods are in points instead of pixels. When setting the frame of an NSWidow you can safely set it in points rather than pixels. 

One caveat to this is that we need to ensure that an NSWindow has a screen when constructed. I suspect that this may be the origin of some bugs currently in libs-gui, but I haven't been able to prove it. 

Added nativeFrame method so we can get the frame of a window in pixels.