Update NSWindow.m
Fixed setMaxSize and setMinSize to ensure that pixels are sent to libs-back. This fixes many of the issues I was noticing with the frame size of windows coming from GORM files being the wrong size. It was capping it by the min and max size which weren't sending to back as pixels, thus down sizing the window.