Update GSDragView.m
Disabled a line of code that is supposed to restrict the window to just the image. This has the side effect of not working on systems where screen scale != 1. On Windowing systems where windows can have transparent backgrounds, this doesn't matter. On WondowMaker; however, this makes the drag window have a black background as WindowMaker doesn't support transparent windows.