Update GSDragView.m
For some reason in the method _clearupWindow, calling [_window setFrame: NSZeroRect display: NO]; causes the window to not be visible when the GSDragView's window needs to be used a second time for a new drag operation. Commenting this line out, fixes that. The line isn't needed as the window will likely be resized for a new drag operation regardless.